fix(export-report-pdf): repoint imports at the renamed settings module (#7221)
