+++ internal-export-file/export-report-pdf/tests/export_report_pdf/test_connector.py
@@ -1,6 +1,6 @@
 import pytest
-from export_report_pdf.settings import ConnectorSettings
 from export_report_pdf.connector import Connector
+from export_report_pdf.settings import ConnectorSettings
 from pycti import OpenCTIConnectorHelper
